requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
joblib==1.4.2
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
